WL#10434: Remote InnoDB Cluster Configuration
With the introduction of the command "SET PERSIST" in MySQL Server
8.0.4 the possibility to remotely persist server configurations becomes
available.
This patch adds support for remote auto-configuration of MySQL Instance
for InnoDB Cluster usage by introducing a new command for such purpose:
dba.configureInstance().
The command must allow remote auto-configuration without any options if
the instance has a version >= 8.0.4. In general, not even a restart of
the instance is required, but for some specific settings, it might be
required. If the instance has a version < 8.0.4, the command won't allow
remote execution, but an optional parameter to specify the configuration
file path can be used.

Additional user-visible changes:

Improved display of messages.
Improved detection of loopback hostnames (for debian/ubuntu) and local
instances from the hostname.
Improved display of required configuration changes.
Replaced auto-configuration of instance on createCluster and addInstance
with a configuration check (checkInstance), aborting the command if the
check fails - Bug#27629803
Option to create @% version of accounts with new password and minimal
grants replaced with option to clone the account into a new one with
custom host specification and same grants and passwords as original.

Additional internal changes:

Started refactoring AdminAPI commands into command objects that
encapsulate undo capabilities and other improvements.
Standardized confirmation prompts.
Standardized AdminAPI option handling and type checking.
Fixed tests to use mocked version of network routines during replay.
Added new tests to cover the functionality of the new command and
updated traces.

Bugs fixed:

BUG#27608299 - Created Admin Account Should Have Same Grants As Original
BUG#27112727 - Creating Cluster On Server With Wrong Binlog_format Fails
BUG#25702994 - NOT NEEDED OUTPUT ON FAILING TO UPDATE THE MYSQL
CONFIGURATION FILE
BUG#27629803 - DBA.CREATECLUSTER AND CLUSTER.ADDINSTANCE SHOULD FAIL IF
INSTANCE NON COMPLIANT

+++ modules/devapi/dynamic_object.cc
@@ -109,7 +109,7 @@ void Dynamic_object::register_dynamic_function(
   DWORD x = 0;
   (void)_BitScanForward(&x, name);
 #else
-  int x = __builtin_ctz(name);
+  size_t x = __builtin_ctz(name);
 #endif
   // We can't register more functions than enabled_paths_ can store.
   assert(static_cast<size_t>(x) < shcore::array_size(enabled_paths_));
